WIP #52, #38: BigTIFF and huge image fixes

+++ Demos/ObjectPascal/LCLImager/mainunit.pas
@@ -261,7 +261,7 @@ type
     FImage: TMultiImage;
     FImageCanvas: TImagingCanvas;
     FFileName: string;
-    FFileSize: Integer;
+    FFileSize: Int64;
     FParam1, FParam2, FParam3: Integer;
     procedure OpenFile(const FileName: string);
     procedure SaveFile(const FileName: string);
